% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vci-package.R
\docType{package}
\name{vci-package}
\alias{vci}
\alias{vci-package}
\title{vci: Vector Control Impact Spatial Modelling}
\description{
Converts spatially varying vector biology, environment, human behaviour, and intervention scenarios into vectorial capacity and vector control impact, the proportional reduction in vectorial capacity attributable to a set of interventions. Implements the Vector Atlas 'Vector Control Impact' framework, computing vector-side transmission potential and its response to control across pixel and administrative-area workflows.
}
\author{
\strong{Maintainer}: Nick Golding \email{nick.golding.research@gmail.com}

Authors:
\itemize{
  \item Nick Golding \email{nick.golding.research@gmail.com}
}

}
\keyword{internal}
